[build] Build all Dune packages in Makefile
We explicitly build all dune packages in Makefile's build target, this
ensures a more stable build setup.

We must do this explicit call as we mark the submodules as
vendored, which makes package building "lazy".

This is the right thing to do, we don't want to build / run
unrelated stuff for vendored packaged like documentation or tests.

We also make the `all` target the default for Make.

I've set `--profile=release` as the build default, this makes
`coq-lsp` plugin loading and .vo file compilation much faster, at the
cost of slower compilation of .ml files.